backend/requirements.txt
# PRISM Web Backend - Python Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication
pyjwt==2.8.0
python-jose[cryptography]==3.3.0

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Environment Variables
python-dotenv==1.0.0

# HTTP Client
httpx==0.26.0

# Utilities
python-multipart==0.0.6

# Logging
loguru==0.7.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
